requirements.txt
asgiref==3.11.0
Django==4.2
django-ckeditor==6.7.3
django-cors-headers==4.9.0
django-jazzmin==2.6.0
django-js-asset==3.1.2
djangorestframework==3.16.1
drf-yasg==1.21.11
inflection==0.5.1
packaging==25.0
pillow==12.1.0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.5
uritemplate==4.2.0
